Simplify SNOMED ECL generation.
It is enormously frustrating that the UK dm+d does not include TF or TFGs. That makes it very difficult to generate ECL expressions from within this library. We can fudge by trying to look for proximal parents, but the best way is to simply provide the right data so that a client, who can combine dm+d and SNOMED CT, can generate the ECL itself.