Add metadata including dm+d release date and version.
This is a breaking change in the backing store, by design. The backing store now incorporates metadata version to ensure compatibility. This means that if we make a breaking change in the database schema we can catch and throw an error rather than some weird, difficult to debug error from deep in the internals.